refactor: reorganize proxy_manager into feature layout
