+++ components/bot_detector/proxy_manager/__init__.py
@@ -1,3 +1,5 @@
-from .core import ProxyManager, Settings
+from .domain.settings import Settings
+from .dtos.proxy import Ports, Proxy
+from .services.manager import ProxyManager
 
-__all__ = ["ProxyManager", "Settings"]
+__all__ = ["ProxyManager", "Settings", "Proxy", "Ports"]
