+++ components/bot_detector/proxy_manager/domain/__init__.py
@@ -1,0 +1,1 @@
+# Domain definitions for proxy management.
