+++ components/bot_detector/proxy_manager/domain/settings.py
@@ -1,0 +1,7 @@
+from pydantic_settings import BaseSettings
+
+
+class Settings(BaseSettings):
+    PROXY_API_KEY: str
+    MAX_CALLS: int = 100
+    INTERVAL: int = 60
