+++ components/bot_detector/proxy_manager/dtos/__init__.py
@@ -1,0 +1,1 @@
+# DTOs for proxy manager component.
