+++ components/bot_detector/proxy_manager/ports/__init__.py
@@ -1,0 +1,1 @@
+# Port definitions for proxy manager component.
