+++ components/bot_detector/proxy_manager/services/__init__.py
@@ -1,0 +1,1 @@
+# Service layer for proxy manager component.
