+++ components/bot_detector/proxy_manager/tests/__init__.py
@@ -1,0 +1,1 @@
+# Placeholder package for component-specific tests (lives under repo test/).
